% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSfrequent}
\alias{BSfrequent}
\title{Daily Buys and Sells}
\format{A matrix with 60 rows and 2 variables:
\describe{
  \item{Buys}{daily buys}
  \item{Sells}{daily sells}
}}
\source{
own simulation
}
\usage{
BSfrequent
}
\description{
A matrix containing three months of synthetic daily buys and sells data representing a frequently traded stock.
}
\keyword{datasets}
