% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSfrequent2015}
\alias{BSfrequent2015}
\title{Daily Buys and Sells}
\format{A matrix with 261 rows and 2 variables:
\describe{
  \item{Buys}{daily buys}
  \item{Sells}{daily sells}
}}
\source{
own simulation
}
\usage{
BSfrequent2015
}
\description{
A matrix containing one year of synthetic daily buys and sells data representing a frequently traded stock.
Rownames equal the business days in 2015.
}
\keyword{datasets}
