% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_post_states.R
\name{ggplot.posterior}
\alias{ggplot.posterior}
\title{Visualization of Posterior Probabilities}
\usage{
\method{ggplot}{posterior}(x)
}
\arguments{
\item{x}{\emph{numeric} matrix returned by \code{\link{posterior}}}
}
\value{
An object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
Visualization of posterior probabilities returned by \code{post_states} with ggplot2
}
\examples{
\dontrun{See Vignette \\code{browseVignette(package = 'pinbasic')}}

}
\seealso{
\code{\link{posterior}}
}
