% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_calc.R
\name{pin_calc}
\alias{pin_calc}
\title{Probability of Informed Trading}
\usage{
pin_calc(param = NULL)
}
\arguments{
\item{param}{\emph{numeric}: (named) vector of model parameters
(valid names: \code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'}),
length must equal 5}
}
\value{
\emph{numeric}: probability of informed trading
}
\description{
Calculates the probability of informed trading.
}
\details{
If names are not set for \code{param} or one or more of the vector names do not match the valid choices, they are internally set to
\code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'} (in this order).
}
\references{
Easley, David et al. (2002) \cr
Is Information Risk a Determinant of Asset Returns? \cr
\emph{The Journal of Finance}, Volume 57, Number 5, pp. 2185 - 2221 \cr
\doi{10.1111/1540-6261.00493}

Easley, David et al. (1996) \cr
Liquidity, Information, and Infrequently Traded Stocks\cr
\emph{The Journal of Finance}, Volume 51, Number 4, pp. 1405 - 1436 \cr
\doi{10.1111/j.1540-6261.1996.tb04074.x}
}
