% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_est.R
\name{pin_est}
\alias{pin_est}
\title{Estimating PIN}
\usage{
pin_est(numbuys = NULL, numsells = NULL, lower = rep(0, 5), upper = c(1,
  1, rep(Inf, 3)))
}
\arguments{
\item{numbuys}{\emph{numeric} vector of daily buys}

\item{numsells}{\emph{numeric} vector of daily sells}

\item{lower}{\emph{numeric} lower bounds for optimization}

\item{upper}{\emph{numeric} upper bounds for optimization}
}
\value{
A list with seven components:
\describe{
\item{Results}{Matrix containing the parameter estimates as well as their standard errors,
 t-values and p-values.}
 \item{ll}{Value of likelihood function returned by \code{nlminb}}
 \item{pin}{Estimated probability of informed trading}
 \item{conv}{Convergence code for nlminb optimization}
 \item{message}{Convergence message returned from the nlminb optimizer}
 \item{iterations}{Number of iterations until convergence of nlminb optimizer}
 \item{init_vals}{Vector of initial values}
 }
}
\description{
Estimates the probability of informed trading (PIN) for daily buys and sells trading data for arbitrary number of trading days.
}
\details{
User-friendly wrapper around workhorse function \code{pin_ll_max}.
\code{\link{nlminb}} function in the \pkg{stats} package is used for maximization.
In the literature, at least data for 60 trading days is recommended to ensure convergence of optimization.
No information about the trading days' dates is needed.
}
\examples{
data("BSfrequent")
res_freq <- pin_est(numbuys = BSfrequent[,"Buys"], numsells = BSfrequent[,"Sells"])
}
\references{
Easley, Hvidkjaer and O'Hara (2002) \cr
Is Information Risk a Determinant of Asset Returns? \cr
\emph{The Journal of Finance}, Volume 57, Number 5, pp. 2185 - 2221

Easley, Kiefer, O'Hara and Paperman (1996) \cr
Liquidity, Information, and Infrequently Traded Stocks\cr
\emph{The Journal of Finance}, Volume 51, Number 4, pp. 1405 - 1436

Easley, David et al. (2010) \cr
Factoring Information into Returns \cr
\emph{Journal of Financial and Quantitative Analysis}, Volume 45, Issue 2, pp. 293 - 309

Ersan, Oguz and Alici, Asli (2016) \cr
An unbiased computation methodology for estimating the probability of informed trading (PIN) \cr
\emph{Journal of International Financial Markets, Institutions and Money}, Volume 43, pp. 74 - 94

Lin, Hsiou-Wei William and Ke, Wen-Chyan (2011) \cr
A computing bias in estimating the probability of informed trading \cr
\emph{Journal of Financial Markets}, Volume 14, Issue 4, pp. 625 - 640

Gan, Quan et al. (2015) \cr
A faster estimation method for the probability of informed trading
using hierarchical agglomerative clustering \cr
\emph{Quantitative Finance}, Volume 15, Issue 11, pp. 1805 - 1821

Yan, Yuxing and Zhang, Shaojun (2012) \cr
An improved estimation method and empirical properties of the probability of informed trading \cr
\emph{Journal of Banking & Finance}, Volume 36, Issue 2, pp. 454 - 467
}
\seealso{
\code{\link{nlminb}},
         \code{\link{initial_vals}}
         \code{\link{pin_est_core}}
         \code{\link{qpin}}
}

