% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_probs.R
\name{posterior}
\alias{posterior}
\title{Posterior Probabilities}
\usage{
posterior(param = NULL, numbuys = NULL, numsells = NULL)
}
\arguments{
\item{param}{\emph{numeric}: (named) vector of model parameters
(valid names: \code{'alpha'}, \code{'delta'}, \code{'epsilon_b'}, \code{'epsilon_s'}, \code{'mu'}),
length must equal 5}

\item{numbuys}{\emph{numeric}: vector of daily buys}

\item{numsells}{\emph{numeric}: vector of daily sells}
}
\value{
\emph{numeric} matrix with columns 'no', 'good' and 'bad' representing posterior probabilities for the corresponding trading day conditions
}
\description{
Calculates posterior probabilities of conditions of trading days
}
\details{
For more details see corresponding section in vignette
}
\examples{
\dontrun{See Vignette \\code{browseVignette(package = 'pinbasic')}}

}
\seealso{
\code{\link{ggplot.posterior}}
}
